# Refactor frequency assertion for improved performance