@thomasckng 2025-07-28
Modified to use `array_equal()` instead of `allclose()`